font_types

Module raw

types for working with raw big-endian bytes

Modules

sealed 🔒

we hide this trait; it isn’t part of the public API, and this clarifies the guarantee that it is only implemented for [u8; N]

Macros

int_scalar 🔒

Structs

BigEndian

A wrapper around raw big-endian bytes for some type.

Traits

FixedSize

A trait for types that have a known, constant size.

Scalar

A trait for font scalars.