font_types

Struct Matrix

#[repr(C)]
pub struct Matrix<T> {
    pub xx: T,
    pub yx: T,
    pub xy: T,
    pub yy: T,
    pub dx: T,
    pub dy: T,
}

A transformation matrix.

This is a column-major 3x2 affine matrix with the following representation:

| xx xy dx |
| yx yy dy |

with the basis vectors (xx, yx), (xy, yy) and the translation (dx, dy).

Transformation of the point (x, y) is applied as follows:

x' = xx * x + xy * y + dx
y' = yx * x + yy * y + dy

Our matrices are typically considered to be right-handed, meaning that positive angles apply counter-clockwise rotations and the rotation matrix for angle a is:

| cos(a) -sin(a) 0 |
| sin(a)  cos(a) 0 |

Matrix multiplication chains transforms in right to left order meaning that for C = A * B, matrix C will apply B and then A.

Fields
xx: Tyx: Txy: Tyy: Tdx: Tdy: T
Implementations

impl<T: Copy> Matrix<T>

pub const fn from_elements(elements: [T; 6]) -> Self

Creates a new matrix from an array of elements in the order [xx, yx, xy, yy, dx, dy].

pub const fn elements(&self) -> [T; 6]

Returns the elements as an array in the order [xx, yx, xy, yy, dx, dy].

pub fn map<U: Copy>(self, f: impl FnMut(T) -> U) -> Matrix<U>

Maps Matrix<T> to Matrix<U> by applying a function to each element.

impl<T: MatrixElement> Matrix<T>

pub const IDENTITY: Self

The identity matrix.

pub fn transform(&self, x: T, y: T) -> (T, T)

Applies the matrix to the given coordinates.