font_types

Struct Point

#[repr(C)]
pub struct Point<T> {
    pub x: T,
    pub y: T,
}

Two dimensional point with a generic coordinate type.

Fields
x: T

X coordinate.

y: T

Y coordinate.

Implementations

impl<T> Point<T>

pub const fn new(x: T, y: T) -> Self

Creates a new point with the given x and y coordinates.

pub const fn broadcast(xy: T) -> Self
where T: Copy,

Creates a new point from a single value assigned to both coordinates.

pub fn map<U>(self, f: impl FnMut(T) -> U) -> Point<U>

Maps Point<T> to Point<U> by applying a function to each coordinate.