Trait font_types::Compatible

pub trait Compatible<Rhs = Self>: Sized {
    // Required method
    fn compatible(&self, other: Rhs) -> bool;
}

A trait for determining whether versions are compatible.

Required Methods

fn compatible(&self, other: Rhs) -> bool

return true if this version is field-compatible with other.

This is kind of poorly defined, but basically means ‘same major version, greater than or equal minor version’.

Object Safety

This trait is not object safe.